% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.iGLLVM.R
\name{biplot.iGLLVM}
\alias{biplot.iGLLVM}
\title{Create biplot from iGLLVM object}
\usage{
\method{biplot}{iGLLVM}(
  object,
  type.posterior.stat = "mean",
  which.lvs = c(1, 2),
  site.labels = F,
  site.col = "black",
  site.cex = 1,
  site.pch = 1,
  spp.col = "blue",
  spp.cex = 1,
  ...
)
}
\arguments{
\item{object}{an object of class 'iGLLVM'}

\item{type.posterior.stat}{which statistic to use from the posterior distributions: mean or mode}

\item{which.lvs}{vector of length two with indexes of LVs to plot}

\item{site.labels}{should sites be plotted as labels or as symbols? Defaults to \code{FALSE} so symbols are plotted}

\item{site.col}{color of plotted site scores}

\item{site.cex}{size of site labels or symbols}

\item{site.pch}{index for site symbols}

\item{spp.col}{color for plotted species scores}

\item{spp.cex}{size of plotted species scores}

\item{...}{other arguments passed to \link{plot}}
}
\description{
Create biplot from iGLLVM object
}
\examples{
\dontrun{
model <- FitGLLVM(matrix(1:10, ncol=5), nLVs=1, family="poisson")
biplot(model)
}
}
\author{
Bert vd Veen
}
