% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFormulaRHS.R
\name{CreateFormulaRHS}
\alias{CreateFormulaRHS}
\title{Create the right hand side of a formula with the specification for the latent variables.}
\usage{
CreateFormulaRHS(
  LVs,
  constrained = FALSE,
  prior.beta = 100,
  hyperprior.LV = NULL,
  init.beta = 0.1
)
}
\arguments{
\item{LVs}{Output from the CreateLVIndices() function}

\item{constrained}{Logical: should the formula be for a constrained model? Defaults to FALSE}

\item{prior.beta}{Prior standard deviation for betas (i.e. column scores). Defaults to 100}

\item{hyperprior.LV}{Hyperparameters for latent variable. Defaults to NULL, otherwise a valid list to be passed to hyper=}

\item{init.beta}{Initial value for betas (i.e. column scores). Defaults to 0.1: cannot be 0 or INLA will complain}
}
\value{
A formula
}
\description{
Create the right hand side of a formula with the specification for the latent variables.
}
\details{
The prior for the betas is currently fixed to be N(0,10^2) (i.e. mean 0, variance 100)
}
\examples{
CreateFormulaRHS(CreateLVIndices(matrix(1:10, ncol=5)))
}
