% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLVIndices.R
\name{CreateLVIndices}
\alias{CreateLVIndices}
\title{Format a data frame/matrix so that it is in the right format for INLA}
\usage{
CreateLVIndices(dat, nLVs = 1)
}
\arguments{
\item{dat}{data. A data frame or matrix}

\item{nLVs}{The number of latent variables required}
}
\value{
A data frame with the same number of columns, but each row only has one value in it (the rest are NAs)
}
\description{
Format a data frame/matrix so that it is in the right format for INLA
}
\examples{
CreateLVIndices(matrix(1:10, ncol=5), nLVs=1)
}
