% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGLLVM.R
\name{FitGLLVM}
\alias{FitGLLVM}
\title{Fits a GLLVM to data}
\usage{
FitGLLVM(Y, X = NULL, nLVs = 1)
}
\arguments{
\item{Y}{A data frame or matrix with the response (assuming counts at the moment)}

\item{X}{A data frame or matrix of covariates (not used yet)}

\item{nLVs}{The number of latent variables required}
}
\value{
A list with fixed, colscores, and roweffs:
the posterior summaries for the fixed effects, the column scores and the row
effects respectively
}
\description{
Fits a GLLVM to data
}
\examples{
FitGLLVM(matrix(1:10, ncol=5), nLVs=1)
}
