% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCovariateData.R
\name{FormatCovariateData}
\alias{FormatCovariateData}
\title{Create the right hand side of a formula with the specification for the latent variables.}
\usage{
FormatCovariateData(
  X = NULL,
  intercept = FALSE,
  AddTerm = NULL,
  intname = NULL,
  nrows = NULL,
  random = NULL
)
}
\arguments{
\item{X}{Covariate data}

\item{intercept}{Logical, should an explicit intercept be added?}

\item{AddTerm}{Term to be added, either a name or a character vector of same length as X. If a single name, the values will be 1:nrow(result)}

\item{intname}{Name of interaction covariate (e.g. if X is supplied, X:column). Defaults to NULL}

\item{nrows}{Number of rows in the result. This gets superceded by anything else (i.e. X or AddTerm) defining the number of rows. Defaults to NULL.}

\item{random}{Vector of names of variables to be made random effects. NULL (the default) if none.}
}
\value{
A data frame with X, an intercept and added term, if desires, and with an attribute that is the part of the formula needed in the model
}
\description{
Create the right hand side of a formula with the specification for the latent variables.
}
\examples{
FormatCovariateData(X=cbind(X1=1:5, X2=2:6), intercept=TRUE, AddTerm = "thing")
}
