% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatDataFrameForLV.R
\name{FormatDataFrameForLV}
\alias{FormatDataFrameForLV}
\title{Format a data frame/matrix so that it is in the right format for INLA}
\usage{
FormatDataFrameForLV(dat)
}
\arguments{
\item{dat}{A data frame or matrix}
}
\value{
A data frame with the same number of columns, but each row only has one value in it (the rest are NAs)
}
\description{
Format a data frame/matrix so that it is in the right format for INLA
}
\examples{
FormatDataFrameForLV(matrix(1:10, ncol=5))
}
