% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forminifera}
\alias{forminifera}
\title{Abundances of Forminifera, along with trait and site data.}
\format{
A list with 4 data frames:
comm: abundances of 24 species (columns) at 31 sites (rows)
traits: data on 3 traits of the 24 species
envir: 10 environmental variables measured at the 31 sites
coord: Coordinates of the 31 sites.

The data are by species and site abundance.
}
\source{
van der Veen (original somewhere else: need to dig out the reference)
}
\usage{
forminifera
}
\description{
Abundances of Forminifera, along with trait and site data.
}
\keyword{datasets}
