% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLVsFromDataFrame.R
\name{MakeLVsFromDataFrame}
\alias{MakeLVsFromDataFrame}
\title{Make a list of indices for latent variables so that it is in the right format for INLA}
\usage{
MakeLVsFromDataFrame(dat, nLVs)
}
\arguments{
\item{dat}{A data frame or matrix}

\item{nLVs}{Number of latent varaibles required}
}
\value{
list of data frames with indices for teh latent variables.
}
\description{
Make a list of indices for latent variables so that it is in the right format for INLA
}
\examples{
MakeLVsFromDataFrame(matrix(1:10, ncol=5), nLVs=1)
}
