% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePriorsInits.R
\name{MakePriorsInits}
\alias{MakePriorsInits}
\title{Create lists of prior specifications and initial values for GLLVM.}
\usage{
MakePriorsInits(
  Y,
  X = NULL,
  W = NULL,
  nLVs = 2,
  RowEff = "fixed",
  ColEff = "fixed",
  RowEffPriorsd = 100,
  ColEffPriorsd = 10,
  PriorLV = NULL
)
}
\arguments{
\item{Y}{A data frame or matrix with the response (assuming counts at the moment)}

\item{X}{A data frame or matrix of covariates (strictly row-level covariates). Can be NULL}

\item{W}{A data frame or matrix of column-level covariates. Can be NULL}

\item{nLVs}{The number of latent variables required}

\item{RowEff}{String indicating what sort of row effect is required. Either none, fixed or random. Defaults to fixed.}

\item{ColEff}{String indicating what sort of column effect is required. Either none, fixed or random. Defaults to fixed.}

\item{RowEffPriorsd}{Prior standard deviation for latent variable, defaults to 100}

\item{ColEffPriorsd}{Prior standard deviation for column scores (the betas for INLA insiders), defaults to 10}

\item{PriorLV}{Hyperprior for the precision of the latent variable, as a list that INLA will understand (sorry). Defaults to NULL, where the default INLA prior will be used}
}
\value{
A data frame with X, an intercept and added term, if desires, and with an attribute that is the part of the formula needed in the model
}
\description{
Create lists of prior specifications and initial values for GLLVM.
}
\examples{
FormatCovariateData(X=cbind(X1=1:5, X2=2:6), intercept=TRUE, AddTerm = "thing")
}
