% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.iGLLVM.R
\name{scores.iGLLVM}
\alias{scores.iGLLVM}
\title{Extract scores from iGLLVM object}
\usage{
\method{scores}{iGLLVM}(object, which = "all", type.posterior.stat = "mean", ...)
}
\arguments{
\item{object}{an object of class 'iGLLVM'}

\item{which}{which scores to retrieve: all, sites, or species}

\item{type.posterior.stat}{which statistic to return of the posterior distributions: mean or mode.}

\item{...}{extra arguments to be passed along.}
}
\description{
Extract scores from iGLLVM object
}
\examples{
\dontrun{
model <- FitGLLVM(matrix(1:10, ncol=5), nLVs=1, family="poisson")
scores(model)
}
}
\author{
Bert vd Veen
}
