% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFormulaRHS.R
\name{CreateFormulaRHS}
\alias{CreateFormulaRHS}
\title{Create the right hand side of a formula with the specification for the latent variables.}
\usage{
CreateFormulaRHS(LVs)
}
\arguments{
\item{LVs}{Output from the CreateLVIndices() function}
}
\value{
A formula
}
\description{
Create the right hand side of a formula with the specification for the latent variables.
}
\examples{
CreateFormulaRHS(CreateLVIndices(matrix(1:10, ncol=5)))
}
