% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGLLVM.R
\name{FitGLLVM}
\alias{FitGLLVM}
\title{Fits a GLLVM to data}
\usage{
FitGLLVM(
  Y,
  X = NULL,
  W = NULL,
  nLVs = 1,
  Family = "gaussian",
  RowEff = "fixed",
  ColEff = "fixed",
  RowEffPriorsd = 100,
  ColEffPriorsd = 100,
  ColScorePriorsd = 10,
  PriorLV = NULL,
  INLAobj = FALSE,
  ...
)
}
\arguments{
\item{Y}{A data frame or matrix with the response (assuming counts at the moment)}

\item{X}{A data frame or matrix of covariates (strictly row-level covariates). Can be NULL}

\item{W}{A data frame or matrix of column-level covariates. Can be NULL}

\item{nLVs}{The number of latent variables required}

\item{Family}{A string indicating the likelihood family. If length 1, it gets repeated with one for each column of the data. For supported distributions see names(inla.models()$likelihood).}

\item{RowEff}{String indicating what sort of row effect is required. Either none, fixed or random. Defaults to fixed.}

\item{ColEff}{String indicating what sort of column effect is required. Either none, fixed or random. Defaults to fixed.}

\item{RowEffPriorsd}{Prior standard deviation for latent variable, defaults to 100}

\item{ColEffPriorsd}{Prior standard deviation for latent variable, defaults to 100}

\item{ColScorePriorsd}{Prior standard deviation for column scores (the betas for INLA insiders), defaults to 10}

\item{PriorLV}{Hyperprior for the precision of the latent variable, as a list that INLA will understand (sorry). Defaults to NULL, where the default INLA prior will be used}

\item{INLAobj}{Should the full INLA object be included in the output object? Defaults to \code{FALSE}}

\item{...}{More arguments to be passed to \code{inla()}}
}
\value{
A list with fixed, rowterm, colterm, colscores, and roweffs, formula, Y, X, family.
the posterior summaries for the fixed effects, row main effects, the column main effect, the column scores and the row
effects respectively. rowterm and colterm may be NULL if they were  not set to be random.
}
\description{
Fits a GLLVM to data
}
\details{
This fits a GLLVM....

Priors for fixed effects can be passed straight to INLA with code like \code{control.fixed = list(mean = ..., prec = ...)}

The column effects have to be fixed effects: if you want them to be random, you have to have fixed rows.
You can just transpose the matrix to get this model.
}
\examples{
FitGLLVM(matrix(1:10, ncol=5), nLVs=1, Family="poisson")
}
